vJtag_11_27_2014_Generic
Generic version of the VHDL for the project. Uses a generic parameter in
the top level to set proper bit sizes. The state machine in the
alu_to_tdo file and the d_logic file will need slight modification to
work for different bit widths, but the project overall is a lot easier
to work with for arbitrary bit widths.